add tool

+++ rust/codeforces/dist.sh
@@ -1,6 +1,8 @@
 #!/bin/zsh
 
-rm -r ./src/contest/*
+if [ "$1"!="n" ]; then
+  rm -r ./src/contest/*
+fi
 cp ./src/main.rs ./src/contest/a.rs
 cp ./src/main.rs ./src/contest/b.rs
 cp ./src/main.rs ./src/contest/c.rs
